Add search without typing the /search command
Now you can just send the bot any text at it will ask you if you want to
search pictures for it.

Here is a example (The box represents a telegram button):

YOU: Austria
BOT: Do you want to search for 'Austria' images?
      --------
     | Search |
      --------

From my usage this is a great improvement, especially for mobile devices
where the slash is not that easy to reach on the keyboard.